+++ lib/modules/student/notifications/event_widget.dart
